simulation output with data grid view

+++ MultiQueueSimulation/MultiQueueModels/Server.cs
@@ -22,7 +22,7 @@ namespace MultiQueueModels
 
         //optional if needed use them
         public int FinishTime { get; set; }
-        public decimal TotalWorkingTime { get; set; }
+        public int TotalWorkingTime { get; set; }
         public decimal Idle { get; set; }
         public decimal Customer { get; set; }
         ///////////// OPERATIONS /////////////
